@@ -8,8 +8,8 @@
     "@types/jest": "^27.5.2",
     "@types/lodash": "^4.14.182",
     "@types/node": "^17.0.41",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.17",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
     "@types/react-router-dom": "^5.3.3",
     "animate.css": "^4.1.1",
     "antd": "^4.21.0",
@@ -17,13 +17,13 @@
     "firebase": "^9.8.3",
     "konva": "^8.3.9",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-firebase-hooks": "^4.0.2",
     "react-hooks-global-state": "^1.0.1",
-    "react-konva": "^17.0.2-6",
+    "react-konva": "^18.1.1",
     "react-router-dom": "6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "react-timer-hook": "^3.0.5",
     "react-use": "^17.4.0",
     "typescript": "^4.7.3",
